Moved zoom logic into a hook (WIP)
